requirements.txt
asgiref==3.5.2
Django==3.2.16
django-environ==0.9.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2==2.9.5
python-decouple==3.7
pytz==2022.6
sqlparse==0.4.3
